Fix warnings

+++ lib/ssd1306_nadyrshin/ssd1306/ssd1306.c
@@ -128,8 +128,8 @@ void SSD1306_SetMemAdressingMode(uint8_t Mode)
 {
   if (Mode > 2)
     return;
-    
-  SendCommand(CMD_SetMemAdressingMode, (uint8_t *)Mode, 1);
+
+  SendCommand(CMD_SetMemAdressingMode, (uint8_t *)(intptr_t)Mode, 1);
 }
 //==============================================================================
